requirements.txt
yfinance
jinja2
requests
openai
pytz
python-dotenv